refactor: extract benchmark stats helpers into shared module
Pulls Stats, computeStats, and printStats out of all three bench files
into example/backend/__tests__/benchmark/stats.ts. printStats now takes
an optional thresholdMs and suppresses the throughput line on
single-iteration runs (where 1/duration is not a meaningful rate).
