+++ example/backend/__tests__/benchmark/fanout.bench.test.ts
@@ -3,6 +3,7 @@ import type { Action } from "keryx";
 import { api, config } from "keryx";
 import { z } from "zod";
 import { HOOK_TIMEOUT, waitFor } from "../setup";
+import { computeStats, printStats } from "./stats";
 
 // ---------------------------------------------------------------------------
 // Stress thresholds
@@ -19,46 +20,6 @@ const STRESS = {
   concurrentFanOuts: 50,
   childrenPerFanOut: 10,
 };
-
-// ---------------------------------------------------------------------------
-// Stats helpers (kept in sync with transports.bench.test.ts)
-// ---------------------------------------------------------------------------
-interface Stats {
-  min: number;
-  avg: number;
-  p50: number;
-  p95: number;
-  p99: number;
-  max: number;
-  rps: number;
-  count: number;
-}
-
-function computeStats(durations: number[]): Stats {
-  const sorted = [...durations].sort((a, b) => a - b);
-  const sum = sorted.reduce((a, b) => a + b, 0);
-  const count = sorted.length;
-  const percentile = (p: number) => sorted[Math.ceil((p / 100) * count) - 1];
-  return {
-    min: sorted[0],
-    avg: Math.round((sum / count) * 100) / 100,
-    p50: percentile(50),
-    p95: percentile(95),
-    p99: percentile(99),
-    max: sorted[count - 1],
-    rps: Math.round((count / (sum / 1000)) * 100) / 100,
-    count,
-  };
-}
-
-function printStats(name: string, stats: Stats) {
-  const fmt = (ms: number) => `${ms.toFixed(2)}ms`;
-  console.log(`\n  Benchmark: ${name}`);
-  console.log(`  Iterations: ${stats.count}`);
-  console.log(
-    `  Latency:  min=${fmt(stats.min)}  avg=${fmt(stats.avg)}  p50=${fmt(stats.p50)}  p95=${fmt(stats.p95)}  p99=${fmt(stats.p99)}  max=${fmt(stats.max)}`,
-  );
-}
 
 // ---------------------------------------------------------------------------
 // Stress actions
